# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/file/upload": {
      "post": {
        "tags": [
          "File"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/file/read/{id}": {
      "get": {
        "tags": [
          "File"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}